move: Header Component to `@/components/Header`
