+++ src/components/Header/index.tsx
@@ -2,10 +2,10 @@ import Link from 'next/link';
 import React from 'react';
 import Image from 'next/image';
 import * as M from '@/styles/Header/HeaderStyle';
-import {HeaderMainDropdown} from '@/components/Layout/Header/HeaderMainDropdown';
-import {HeaderCalcDropdown} from '@/components/Layout/Header/HeaderCalcDropdown';
-import {HeaderIntroDropdown} from '@/components/Layout/Header/HeaderIntroDropdown';
-import {HeaderBoardDropdown} from '@/components/Layout/Header/HeaderBoardDropdown';
+import {HeaderMainDropdown} from '@/components/Header/HeaderMainDropdown';
+import {HeaderCalcDropdown} from '@/components/Header/HeaderCalcDropdown';
+import {HeaderIntroDropdown} from '@/components/Header/HeaderIntroDropdown';
+import {HeaderBoardDropdown} from '@/components/Header/HeaderBoardDropdown';
 import {MobileMenuTab} from './MobileMenuTab';
 
 import {useState} from 'react';
